xxhash transition hashbuffer fix
For transition adding/fetching, a buffer is passed to the
XXH3_64bits() function. This function also takes another
argument representing the size of said buffer. Before this
commit, the number of buffer elements was passed to this
function. Now the total number of bytes in the buffer is
passed, which should greatly improve performance of the
import and run process, especially for machines with large
alphabets and/or many transitions.